closeAndMoveFile()  function
Implemented a helper function that can be used when File I/O operations go poorly.
Argument: The file object with the opened file
Body: Closes the file, stores its contents, and writes those contents to a file with [OLD] in name. If the file is the service Information file for the program, delete the file (so that it can be automatically re-made the next time the program is executed).